+++ level/views.py
@@ -44,6 +44,10 @@ def add_level(request):
 
 @login_required(login_url='login')
 def del_level(request, pk):
-    get_level = Level.objects.get(pk=pk)
-    get_level.delete()
+    if request.user.is_staff:
+        get_level = Level.objects.get(pk=pk)
+        get_level.delete()
+        messages.success(request, 'تم الحذف بنجاح')
+        return redirect('level')
+    messages.warning(request, 'ليس لديك صلاحية للقيام بهذه العملية')
     return redirect('level')
